+++ src/model/football/FootballAttributes.java
@@ -26,7 +26,7 @@ public class FootballAttributes
 
 	public int getGoa()
 	{
-		return attributes.get(0).getRating();
+		return (int)attributes.get(0).getRating();
 	}
 
 	public void setGoa(int goa)
@@ -38,7 +38,7 @@ public class FootballAttributes
 
 	public int getDef()
 	{
-		return attributes.get(1).getRating();
+		return (int)attributes.get(1).getRating();
 	}
 
 	public void setDef(int def)
@@ -50,7 +50,7 @@ public class FootballAttributes
 
 	public int getMid()
 	{
-		return attributes.get(2).getRating();
+		return (int)attributes.get(2).getRating();
 	}
 
 	public void setMid(int mid)
@@ -62,7 +62,7 @@ public class FootballAttributes
 
 	public int getOff()
 	{
-		return attributes.get(3).getRating();
+		return (int)attributes.get(3).getRating();
 	}
 
 	public void setOff(int off)
@@ -74,7 +74,7 @@ public class FootballAttributes
 
 	public int getSho()
 	{
-		return attributes.get(4).getRating();
+		return (int)attributes.get(4).getRating();
 	}
 
 	public void setSho(int sho)
@@ -86,7 +86,7 @@ public class FootballAttributes
 
 	public int getPas()
 	{
-		return attributes.get(5).getRating();
+		return (int)attributes.get(5).getRating();
 	}
 
 	public void setPas(int pas)
@@ -98,7 +98,7 @@ public class FootballAttributes
 
 	public int getTec()
 	{
-		return attributes.get(6).getRating();
+		return (int)attributes.get(6).getRating();
 	}
 
 	public void setTec(int tec)
@@ -110,7 +110,7 @@ public class FootballAttributes
 
 	public int getSpe()
 	{
-		return attributes.get(7).getRating();
+		return (int)attributes.get(7).getRating();
 	}
 
 	public void setSpe(int spe)
@@ -122,7 +122,7 @@ public class FootballAttributes
 
 	public int getHea()
 	{
-		return attributes.get(8).getRating();
+		return (int)attributes.get(8).getRating();
 	}
 
 	public void setHea(int hea)
